add manage event function for organzations
