+++ capstone/src/main/resources/static/event.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="event.css">
   </head>
   
-  <body onload="getEvents()">
+  <body onload="getEvents();loadEventInfo();">
 
     <div class="menu">
       <nav>
