+++ capstone/src/main/resources/static/event.js
@@ -31,3 +31,18 @@ async function newEvent() {
   await fetch('save-event', {method: 'POST'});
 }
 
+function loadEventInfo() {
+  const event = window.location.hash.substring(1);
+  if (event != "") {
+    fetch('get-event?event-id=' + event).then(response => response.json()).then((data) => {
+      document.getElementById("eventTitle").value = data.eventTitle;
+      document.getElementById("eventDateTime").value = data.eventDateTime;
+      document.getElementById("eventLatitude").value = data.eventLatitude;
+      document.getElementById("eventLongitude").value = data.eventLongitude;
+      document.getElementById("eventDescription").value = data.description;
+
+    });
+  }
+
+}
+
