+++ capstone/src/main/resources/static/manageevents.html
@@ -4,7 +4,7 @@
     <meta>
     <title>My Account</title>
     <link rel="stylesheet" href="style.css">
-    <script src="script.js"></script>
+    <script src="user.js"></script>
 </head>
 <body onload="getOrganizationEvents();">
     <nav class="nav-bar">
@@ -18,8 +18,8 @@
     <br/>
     <div class="left nav-bar" id="org-nav">
         <ul>
-            <li class="active"><a href="profile.html">My Profile</a></li>
-            <li><a href="manageevents.html">Manage my Events</a></li>
+            <li><a href="profile.html">My Profile</a></li>
+            <li class="active"><a href="manageevents.html">Manage my Events</a></li>
             <li><a href="event.html">Create an Event</a></li>
         </ul>
     </div>
