+++ capstone/src/main/resources/static/savedevents.html
@@ -4,7 +4,7 @@
     <meta>
     <title>My Account</title>
     <link rel="stylesheet" href="style.css">
-    <script src="script.js"></script>
+    <script src="user.js"></script>
 </head>
 <body onload="getIndividualEvents();">
     <nav class="nav-bar">
