+++ capstone/src/main/resources/static/style.css
@@ -109,7 +109,7 @@ input[type=submit]:hover {
   margin-right: auto;
 }
 
-#organization, #no-profile, #user, #org-nav {
+#organization, #no-profile, #user {
   display: none;
 }
